switch copies double FK from IdClass to Embedded
