+++ app/src/main/java/main/WTLibraryApp/IWTLibCopyRepository.java
@@ -6,6 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IWTLibCopyRepository extends JpaRepository<Copies, Long>{
-	Optional<Copies> findByBookIdCopyId(long bookId, long copyId);
+public interface IWTLibCopyRepository extends JpaRepository<Copies, CopiesId>{
+	Optional<Copies> findByCopiesId(CopiesId copiesId);
+
+	Optional<Copies> deleteByCopiesId(CopiesId copiesId);
 }
